smartgov-ai/backend/requirements.txt
# Django & Django REST Framework Core
Django==5.2.8
djangorestframework==3.14.0
djangorestframework-authtoken==3.14.0
drf-spectacular==0.27.1
django-cors-headers==4.3.1

# Environment & Configuration
python-decouple==3.8
python-dotenv==1.0.0
django-environ==0.21.0

# AI & LLM Integration
openai==1.3.0
requests==2.31.0

# Database & ORM
psycopg2-binary==2.9.9
django-extensions==3.2.3

# File Handling & Storage
Pillow==10.1.0
django-storages==1.14.2

# Task Queue & Caching
celery==5.3.4
redis==5.0.1
django-redis==5.4.0

# Production Web Server
gunicorn==21.2.0
whitenoise==6.6.0

# Security & Authentication
cryptography==41.0.7

# Development & Testing
django-debug-toolbar==4.2.0
black==23.11.0
flake8==6.1.0
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
slugify==0.0.1

# Monitoring & Error Tracking
sentry-sdk==1.38.0
newrelic==9.3.0

# Email Backend
django-sendgrid-v5==1.2.2

# API Filtering & Pagination
django-filter==23.4

# Rate Limiting
django-ratelimit==4.1.0

# Real-time Features (Optional)
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0

# Admin Enhancement
django-admin-interface==0.25.0

# Localization & Translation
django-modeltranslation==0.18.11

# Code Quality & Formatting
isort==5.13.2
flake8-docstrings==1.7.0
